bot-api/requirements.txt
fastapi
uvicorn
httpx
requests
PyMuPDF
python-multipart
easyocr
numpy<2
Pillow
psutil